Complete BeaconPipeline. Bunch of Channel fixes. Create and pass BeaconPipelineChannelTest

+++ wire/src/main/java/org/ethereum/beacon/wire/message/RequestMessagePayload.java
@@ -1,8 +1,6 @@
 package org.ethereum.beacon.wire.message;
 
-import tech.pegasys.artemis.util.uint.UInt64;
-
 public abstract class RequestMessagePayload extends MessagePayload {
 
-  public abstract UInt64 getMethodId();
+  public abstract int getMethodId();
 }
